requirements.txt
# Web Development
Flask==3.0.2
flask-cors==4.0.0
requests==2.31.0

# AI and Text Processing
google-generativeai==0.3.2
beautifulsoup4==4.12.3

# Environment Variables
python-dotenv==1.0.1

# Development Tools
black==24.2.0
flake8==7.0.0
